ci: build docs as part of build check (#52)
* ci: build docs as part of build check

This ensures that part of the publish step is checked in feature
branches and won't break on `master` without warning.

* Revert update to typescript

This is causing failures with build-docs command. There is a PR #51 that
will update Webpack to v5 and ts-loader to work with the newer version.
This will allow using the latest TypeScript version.

* Don't run build-docs on Node 18

Need to investigate issue and find a fix but for now just not running on
this version.